pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.6",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "plotly>=6.1.2",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "streamlit>=1.45.1",
    "xlrd>=2.0.1",
]
